+++ test/acceptance.js
@@ -1,6 +1,6 @@
 var request = require('supertest');
 var path = require('path');
-var gibRun = require('..').start({
+var gibRunss = require('..').start({
 	root: path.join(__dirname, "data"),
 	port: 0,
 	open: false
@@ -8,35 +8,35 @@ var gibRun = require('..').start({
 
 describe('basic functional tests', function(){
 	it('should respond with index.html', function(done){
-		request(gibRun)
+		request(gibRunss)
 			.get('/')
 			.expect('Content-Type', 'text/html; charset=UTF-8')
 			.expect(/hello world/i)
 			.expect(200, done);
 	});
 	it('should have injected script', function(done){
-		request(gibRun)
+		request(gibRunss)
 			.get('/')
 			.expect('Content-Type', 'text/html; charset=UTF-8')
 			.expect(/<script [^]+?live reload enabled[^]+?<\/script>/i)
 			.expect(200, done);
 	});
 	it('should inject script when tags are in CAPS', function(done){
-		request(gibRun)
+		request(gibRunss)
 			.get('/index-caps.htm')
 			.expect('Content-Type', 'text/html; charset=UTF-8')
 			.expect(/<script [^]+?live reload enabled[^]+?<\/script>/i)
 			.expect(200, done);
 	});
 	it('should inject to <head> when no <body>', function(done){
-		request(gibRun)
+		request(gibRunss)
 			.get('/index-head.html')
 			.expect('Content-Type', 'text/html; charset=UTF-8')
 			.expect(/<script [^]+?live reload enabled[^]+?<\/script>/i)
 			.expect(200, done);
 	});
 	it('should inject also svg files', function(done){
-		request(gibRun)
+		request(gibRunss)
 			.get('/test.svg')
 			.expect('Content-Type', 'image/svg+xml')
 			.expect(function(res) {
@@ -46,7 +46,7 @@ describe('basic functional tests', function(){
 			.expect(200, done);
 	});
 	it('should not inject html fragments', function(done){
-		request(gibRun)
+		request(gibRunss)
 			.get('/fragment.html')
 			.expect('Content-Type', 'text/html; charset=UTF-8')
 			.expect(function(res) {