+++ test/mount.js
@@ -1,6 +1,6 @@
 var request = require('supertest');
 var path = require('path');
-var gibRun = require('..').start({
+var gibRuns = require('..').start({
 	root: path.join(__dirname, "data"),
 	port: 0,
 	open: false,
@@ -12,14 +12,14 @@ var gibRun = require('..').start({
 
 describe('mount tests', function() {
 	it('should respond with sub.html', function(done) {
-		request(gibRun)
+		request(gibRuns)
 			.get('/mounted/sub.html')
 			.expect('Content-Type', 'text/html; charset=UTF-8')
 			.expect(/Subdirectory/i)
 			.expect(200, done);
 	});
 	it('should respond with style.css', function(done) {
-		request(gibRun)
+		request(gibRuns)
 			.get('/style')
 			.expect('Content-Type', 'text/css; charset=UTF-8')
 			.expect(/color/i)